requirements.txt
pandas
numpy
streamlit
streamlit-option-menu
mysql-connector-python

-e .